+++ tests/list.spec.js
@@ -35,7 +35,7 @@ describe('lst', function () {
 
     it('should invoke', function() {
         var list = pyscript.list([1, 2, 3, 4]);
-        expect(list.invoke(function(a) {return a+1})).toEqual([2, 3, 4, 5]);
+        expect(list.map(function(a) {return a+1})).toEqual([2, 3, 4, 5]);
     });